requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

SQLAlchemy>=2.0.0
asyncpg>=0.29.0

pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

Jinja2>=3.1.0
python-multipart>=0.0.9

pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
